fix(network): treat null reload state as enabled on destroy
State written before `reload` existed carries no value for it, and with no
schema version bump the framework decodes that as null. Delete read it as
false, so removing a pre-upgrade interface staged the deletion and never
reloaded, leaving the interface live with no diagnostic.

Backfill null to true on read, and treat null as enabled in Delete. The read
backfill also removes the spurious `reload = true` in-place update, and the
`ifreload -a` it carries, that every existing resource would otherwise get on
the first plan after upgrading.

Cover it with staged-destroy tests on all three resources: the node must still
report pending changes once the harness has destroyed with `reload = false`.
A staged delete drops out of the interface list immediately, so `active`
cannot tell it apart from an applied one.

Also document that any reload on a node applies every staged change on it, and
move the reload tests into their resource test files with the shared
assertions in a helper.
